+++ core/src/test/java/org/hzt/utils/sequences/WindowedSequenceTest.java
@@ -4,6 +4,7 @@ import org.hzt.utils.collections.ListX;
 import org.hzt.utils.iterables.primitives.IntNumerable;
 import org.hzt.utils.ranges.IntRange;
 import org.hzt.utils.tuples.Pair;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,6 +13,7 @@ import java.time.LocalDate;
 import java.time.Month;
 import java.util.function.IntSupplier;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 class WindowedSequenceTest {
@@ -59,7 +61,7 @@ class WindowedSequenceTest {
                 .mapToLong(ListX::size)
                 .toArray();
 
-        assertArrayEquals(new long[] {1, 3, 5, 8, 10, 12, 14, 15, 15, 15, 14, 13, 11, 9, 6, 4, 2, 1, 1, 1}, chunkSizes);
+        assertArrayEquals(new long[]{1, 3, 5, 8, 10, 12, 14, 15, 15, 15, 14, 13, 11, 9, 6, 4, 2, 1, 1, 1}, chunkSizes);
     }
 
     private IntSupplier sineWaveGenerator() {
@@ -88,5 +90,36 @@ class WindowedSequenceTest {
                 () -> assertEquals(ListX.of(4, 5, 6, 7, 8), windows.first()),
                 () -> assertEquals(ListX.of(49, 50), windows.last())
         );
+    }
+
+    @Nested
+    class NextChunkIf {
+
+        @Test
+        void testNextChunkIf() {
+            final var chunks = Sequence.of(1, 2, 3, 4, 5, 6, 2, 4, 5, 123, 2, 3, 1)
+                    .nextChunkedIf(i -> i == 3)
+                    .toList();
+
+            assertThat(chunks).containsExactly(
+                    ListX.of(1, 2),
+                    ListX.of(3, 4, 5, 6, 2, 4, 5, 123, 2),
+                    ListX.of(3, 1)
+            );
+        }
+
+        @Test
+        void testNextChunkIfStartWithPredicate() {
+            final var chunks = Sequence.of(3, 1, 2, 3, 4, 5, 6, 2, 4, 5, 123, 2, 3)
+                    .nextChunkedIf(i -> i == 3)
+                    .toList();
+
+            assertThat(chunks).containsExactly(
+                    ListX.of(3, 1, 2),
+                    ListX.of(3, 4, 5, 6, 2, 4, 5, 123, 2),
+                    ListX.of(3)
+            );
+        }
+
     }
 }